fix(android): fix delay() Int to Long conversion
delay() from kotlinx.coroutines takes a Long parameter. The delayMs
variable is Int, so it needs .toLong() conversion.

+++ Android_app/SetupWizard/app/src/main/java/com/testplaybyte/setupwizard/SetupWizardApp.kt
@@ -376,7 +376,7 @@ private fun CountUpText(
 ) {
     val animatedValue = remember(target) { Animatable(0f) }
     LaunchedEffect(target) {
-        delay(delayMs)
+        delay(delayMs.toLong())
         animatedValue.animateTo(
             target.toFloat(),
             tween(durationMs, easing = FastOutSlowInEasing),
